Dixon Price function implementation

+++ GeneticAlgorithm/src/main/java/com/cefetmg/geneticalgorithm/commom/individual/Individual.java
@@ -10,6 +10,6 @@ public interface Individual {
     
     public Double getEvaluation();
     
-    public List<Integer> getGenes();
+    public List<Double> getGenes();
     
 }
